robot kinda works now
Need to integrate everything togehter

- as well as timings for going up and down

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/command/ArmDownCommand.java
@@ -8,10 +8,10 @@ import org.firstinspires.ftc.teamcode.subsystem.ArmSubsystem;
 public class ArmDownCommand extends CommandBase {
     ArmSubsystem arm;
     int slide, time;
-    int armPos;
+    double armPos;
     double wrist;
     ElapsedTime timer;
-    public ArmDownCommand(ArmSubsystem arm, int slide, int armPos, double wrist, int time) {
+    public ArmDownCommand(ArmSubsystem arm, int slide, double armPos, double wrist, int time) {
         this.arm = arm;
         this.slide = slide;
         this.armPos = armPos;
