+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/command/ArmUpCommand.java
@@ -8,10 +8,10 @@ import org.firstinspires.ftc.teamcode.subsystem.ArmSubsystem;
 public class ArmUpCommand extends CommandBase {
     ArmSubsystem arm;
     int slide, time;
-    int armPos;
+    double armPos;
     double wrist;
     ElapsedTime timer;
-    public ArmUpCommand(ArmSubsystem arm, int slide, int armPos, double wrist, int time) {
+    public ArmUpCommand(ArmSubsystem arm, int slide, double armPos, double wrist, int time) {
         this.arm = arm;
         this.slide = slide;
         this.armPos = armPos;
